'Example_formatting_done_to_each_V8_tissue_file'
This file provides an example of the pre-formatting I did to each of the v8 eGene eQTL GTex files. It shows the directory structure in ./v8 where I put each of the tissue folders. Instead of uploading each script for the 49 tissues, this example shows what I did for breast. Each of the formatted tissue files were later merged.